// src/lib/types/db/dfa/dfaParty.ts
export interface DfaPartyDbModel {
	RunID: string;
	P1PlayerID: string;
	P2PlayerID: string;
	P3PlayerID: string;
	P4PlayerID: string;
	P5PlayerID: string;
	P6PlayerID: string;
	P7PlayerID: string;
	P8PlayerID: string;
	P1RunCharacter: string;
	P2RunCharacter: string;
	P3RunCharacter: string;
	P4RunCharacter: string;
	P5RunCharacter: string;
	P6RunCharacter: string;
	P7RunCharacter: string;
	P8RunCharacter: string;
	Patch: string;
	Drill: string;
	Buff: string;
	Time: string;
	P1MainClass: string;
	P2MainClass: string;
	P3MainClass: string;
	P4MainClass: string;
	P5MainClass: string;
	P6MainClass: string;
	P7MainClass: string;
	P8MainClass: string;
	P1SubClass: string;
	P2SubClass: string;
	P3SubClass: string;
	P4SubClass: string;
	P5SubClass: string;
	P6SubClass: string;
	P7SubClass: string;
	P8SubClass: string;
	PartySize: string;
	P1Link: string;
	P2Link: string;
	P3Link: string;
	P4Link: string;
	P5Link: string;
	P6Link: string;
	P7Link: string;
	P8Link: string;
	Notes: string;
	SubmissionTime: string;
	SubmitterID: string;
	ServerID: string;
	P1PlayerName: string;
	P1PlayerCName: string;
	P1PlayerNameType: string;
	P1PlayerNameColor1: string;
	P1PlayerNameColor2: string;
	P1PlayerServer: string;
	P1PlayerPrefN: string;
	P2PlayerName: string;
	P2PlayerCName: string;
	P2PlayerNameType: string;
	P2PlayerNameColor1: string;
	P2PlayerNameColor2: string;
	P2PlayerServer: string;
	P2PlayerPrefN: string;
	P3PlayerName: string;
	P3PlayerCName: string;
	P3PlayerNameType: string;
	P3PlayerNameColor1: string;
	P3PlayerNameColor2: string;
	P3PlayerServer: string;
	P3PlayerPrefN: string;
	P4PlayerName: string;
	P4PlayerCName: string;
	P4PlayerNameType: string;
	P4PlayerNameColor1: string;
	P4PlayerNameColor2: string;
	P4PlayerServer: string;
	P4PlayerPrefN: string;
	P5PlayerName: string;
	P5PlayerCName: string;
	P5PlayerNameType: string;
	P5PlayerNameColor1: string;
	P5PlayerNameColor2: string;
	P5PlayerServer: string;
	P5PlayerPrefN: string;
	P6PlayerName: string;
	P6PlayerCName: string;
	P6PlayerNameType: string;
	P6PlayerNameColor1: string;
	P6PlayerNameColor2: string;
	P6PlayerServer: string;
	P6PlayerPrefN: string;
	P7PlayerName: string;
	P7PlayerCName: string;
	P7PlayerNameType: string;
	P7PlayerNameColor1: string;
	P7PlayerNameColor2: string;
	P7PlayerServer: string;
	P7PlayerPrefN: string;
	P8PlayerName: string;
	P8PlayerCName: string;
	P8PlayerNameType: string;
	P8PlayerNameColor1: string;
	P8PlayerNameColor2: string;
	P8PlayerServer: string;
	P8PlayerPrefN: string;
	SubmitterName: string;
	SubmitterCName: string;
	SubmitterNameType: string;
	SubmitterNameColor1: string;
	SubmitterNameColor2: string;
	SubmitterPrefN: string;
}