// src/lib/types/db/dfa/dfaSolo.ts
export interface DfaSoloDbModel {
	RunID: string;
	PlayerID: string;
	RunCharacter: string;
	Patch: string;
	Drill: string;
	Support: string;
	Time: string;
	MainClass: string;
	SubClass: string;
	W1: string;
	W2: string;
	W3: string;
	W4: string;
	W5: string;
	W6: string;
	Link: string;
	Notes: string;
	SubmissionTime: string;
	SubmitterID: string;
	PlayerName: string;
	PlayerCName: string;
	PlayerNameType: string;
	PlayerNameColor1: string;
	PlayerNameColor2: string;
	PlayerServer: string;
	PlayerPrefN: string;
	SubmitterName: string;
	SubmitterCName: string;
	SubmitterNameType: string;
	SubmitterNameColor1: string;
	SubmitterNameColor2: string;
	SubmitterPrefN: string;
}